move controller into teacher namespace, remove locale param
